Major rework to use new core decoder on Pro Mini
Change of hardware and change of approach in algorithm. Have now
used logic analyzer to develop code that seems to more reliably
decode the signal.

Relying on a bunch of callbacks to generalise the same code for
use in Saleae Logic software and on (soon) a Pro Mini.

Still lots of rough edges here but is building in IDE.